Conflicts:
	.travis.yml
	Library/Formula/android-sdk.rb
	Library/Formula/app-engine-go-64.rb
	Library/Formula/cairo.rb
	Library/Formula/go.rb
	Library/Formula/libbluray.rb
	Library/Formula/newt.rb
	Library/Formula/poco.rb
	Library/Formula/rust.rb
	Library/Formula/subversion.rb
	Library/Homebrew/bottles.rb
	Library/Homebrew/cmd/config.rb
	Library/Homebrew/cmd/doctor.rb
	Library/Homebrew/cmd/test-bot.rb
	Library/Homebrew/extend/ENV/shared.rb
	Library/Homebrew/formula_support.rb
	Library/Homebrew/os/mac.rb
	README.md

+++ Library/Homebrew/keg.rb
@@ -186,7 +186,7 @@ class Keg
     remove_oldname_opt_record
   end
 
-  def unlink
+  def unlink(mode = OpenStruct.new)
     ObserverPathnameExtension.reset_counts!
 
     dirs = []
@@ -201,6 +201,12 @@ class Keg
 
         # check whether the file to be unlinked is from the current keg first
         if dst.symlink? && src == dst.resolved_path
+          if mode.dry_run
+            puts dst
+            Find.prune if src.directory?
+            next
+          end
+
           dst.uninstall_info if dst.to_s =~ INFOFILE_RX
           dst.unlink
           Find.prune if src.directory?
@@ -208,15 +214,22 @@ class Keg
       end
     end
 
-    remove_linked_keg_record if linked?
-
-    dirs.reverse_each(&:rmdir_if_possible)
+    unless mode.dry_run
+      remove_linked_keg_record if linked?
+      dirs.reverse_each(&:rmdir_if_possible)
+    end
 
     ObserverPathnameExtension.total
   end
 
   def lock
-    FormulaLock.new(name).with_lock { yield }
+    FormulaLock.new(name).with_lock do
+      if oldname_opt_record
+        FormulaLock.new(oldname_opt_record.basename.to_s).with_lock { yield }
+      else
+        yield
+      end
+    end
   end
 
   def completion_installed?(shell)
@@ -303,7 +316,9 @@ class Keg
       when "dtrace" then :mkpath
       when /^gdk-pixbuf/ then :mkpath
       when "ghc" then :mkpath
+      when /^gio/ then :mkpath
       when "lua" then :mkpath
+      when /^mecab/ then :mkpath
       when /^node/ then :mkpath
       when /^ocaml/ then :mkpath
       when /^perl5/ then :mkpath
